requirements.txt
tensorflow
numpy
pillow
keras
Flask